feat: secure routes with temporary validation
As the prototype should only allow certain user accounts to access the
backend, I added a hardcoded values for the discord users.

This is by far not wanted but the quickest solution for a some basic
security of other people that would otherwise be able to
update/delete/insert data in our useable prototype which should not be
the case.

Later on this will be replaced with a general user account which has
only access to specific parts of the data. But this is for now not in
the scope of a first mvp.