style(cost): update style and return of cost endpoint
Generally all my endpoints are returning ids only. This is for things
like account. So instead of returning the account object the user needs
to get the account later on to know what data exists.

I added the return of name for convenience but it seems like that it
does not make a difference for the frontend, so I reverted it to keep in
line with the other implementations.

After looking into sqlx there is no easy way of automatically adding
joins, because of this i am guessing this will stay like this (with ids)
for now.

This is relevant for issue #3.