fix(num-conv): round floats after multiply of 100
This is needed as otherwise some examples will result in too low of a
value. For example 17.40 * 100 will result in 1738.99999 which then
would be 1738 int instead of the expected 1740. For that we need to
round it before (might need to do something else, but for now that
should be fine.

As I do that in more places, I put it in its own helper function for
now. So changing it later should be easier as well.